requirements.txt
# GUI Framework
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0

# AI/LLM Integration
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0
langchain>=0.1.0
langgraph>=0.0.20

# Document Processing
python-docx>=1.1.0
pypandoc>=1.12
ebooklib>=0.18
markdown>=3.5

# Spell Check and Grammar
language-tool-python>=2.8.0
pyspellchecker>=0.7.2

# Image Processing
Pillow>=10.2.0
requests>=2.31.0

# Data Management
pydantic>=2.5.0
pyyaml>=6.0.1

# Utilities
python-dotenv>=1.0.0